requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
PyJWT==2.8.0
ShopifyAPI>=12.0.0,<13.0.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.7
